+++ ICD.Profound.ConnectPRO/Themes/UserInterface/Presenters/VideoConference/VtcBasePresenter.cs
@@ -28,8 +28,6 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.VideoConferenc
 	[PresenterBinding(typeof(IVtcBasePresenter))]
 	public sealed class VtcBasePresenter : AbstractPopupPresenter<IVtcBaseView>, IVtcBasePresenter
 	{
-		public event EventHandler OnActiveConferenceControlChanged;
-
 		private readonly IVtcCallListTogglePresenter m_CallListTogglePresenter;
 		private readonly IVtcContactsNormalPresenter m_ContactsNormalPresenter;
 		private readonly IVtcContactsPolycomPresenter m_ContactsPolycomPresenter;
@@ -388,6 +386,9 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.VideoConferenc
 					if (active != null)
 						active.Hangup();
 				}
+
+				if (Room != null)
+					Room.FocusSource = null;
 			}
 
 			UpdateCodecAwakeState(true);