+++ ICD.Profound.ConnectPRO/Themes/UserInterface/Presenters/WebConference/WtcBasePresenter.cs
@@ -333,6 +333,9 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.WebConference
 					if (active != null)
 						active.Hangup();
 				}
+
+				if (Room != null)
+					Room.FocusSource = null;
 			}
 
 			UpdateCodecAwakeState(true);